+++ src/modules/payment/stripeWebhook.ts
@@ -345,7 +345,7 @@ const stripeWebhook = catchAsync(async (req: Request, res: Response) => {
                     ? durationHours.toString()
                     : '',
                   deliveryDate: updatedDeliveryDate,
-                  totalPrice: (parseFloat(orderData?.totalPrice as string) || 0) + (updateMessage.amount || 0).toString(),
+                  totalPrice: (parseInt(orderData?.totalPrice as string) || 0) + (updateMessage.amount || 0).toString(),
                 },
               });
             } else {
